Experimenting with client-side code
Lots of weirdness going on. Fixed some bugs with the objective tracking,
that all should be working now. PurchaseInteraction.onPurchase only
fires server-side, so may have to try to poll for it client-side. This
may have some severe limitations, so I have to decide if it's even worth
it to check for things client-side.