Major networking refactor
Lots of changes! Basically everything was redone. Client vs server
behavior should be much more consistent and anticipated. Checks are made
for whether the mod is running client-only or as a server as well. Some
better code design is also being used, using factories to instantiate
components dynamically depending on circumstances.

InteractableObjectiveController and
MultishopInteractableObjectiveController may not work yet. They have
code that requires events that only run server-side, so we may have to
write some server wrapper code to handle it, but gonna take a bit to
think about that.